Moved more code to compilation module
